Show failure details for put saves
Tiddlyhost puts plain text error message in the response body. It's
not a Webdav compatible thing (the spec requires XML I think), but
it works fine for what we're doing on Tiddlyhost.

This change lets siteleteer users see the error message, for example
the "If this is your site please log in..." message for a 403
response.

There's a little extra work needed since res.text() is a promise.
The technique is (once again) based on similar code in Feather Wiki.